extracted interface from resolver strategy implementation
